RFC: clang-tidy: Enable bugprone-integer-division
Should this be enabled?
Pro: It seems to catch real errors. Makes code more readable.
     (I'm not sure I have fixed all the correct way which means
      the code was hard to read before)
Cons: It has some false postives

If so how do we want to mark attentional int divisions?
(int)(a/b)            #in this patch
static_cast<int>(a/b) #style guide but kind of long for this purpose
int_div(a,b)          #new function for this purpose

+++ src/gromacs/domdec/domdec_topology.cpp
@@ -1212,7 +1212,7 @@ static void combine_idef(t_idef *dest, const thread_work_t *src, int nsrc,
                 if ((ild->nr + n)/nral1 > vsite->vsite_pbc_loc_nalloc[ftv])
                 {
                     vsite->vsite_pbc_loc_nalloc[ftv] =
-                        over_alloc_large((ild->nr + n)/nral1);
+                        over_alloc_large((int)((ild->nr + n)/nral1));
                     srenew(vsite->vsite_pbc_loc[ftv],
                            vsite->vsite_pbc_loc_nalloc[ftv]);
                 }
